requirements.txt
streamlit==1.29.0
openai==1.54.0
qdrant-client==1.7.0
python-dotenv==1.0.0
